feat: move Puck to /edit/* to improve caching
